requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
aniso8601==8.0.0
    # via
    #   -r requirements.in
    #   flask-restful
attrs==19.3.0
    # via -r requirements.in
blinker==1.4
    # via flask-mail
certifi==2019.11.28
    # via
    #   -r requirements.in
    #   requests
chardet==3.0.4
    # via
    #   -r requirements.in
    #   requests
click==7.1.1
    # via
    #   -r requirements.in
    #   flask
    #   pip-tools
cycler==0.10.0
    # via matplotlib
decorator==4.4.2
    # via networkx
fdict==0.8.1
    # via -r requirements.in
flask==1.1.1
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-mail
    #   flask-marshmallow
    #   flask-restful
flask-cors==3.0.8
    # via -r requirements.in
flask-mail==0.9.1
    # via -r requirements.in
flask-marshmallow==0.11.0
    # via -r requirements.in
flask-restful==0.3.8
    # via -r requirements.in
future==0.18.2
    # via pygam
gevent==20.6.2
    # via -r requirements.in
greenlet==0.4.16
    # via
    #   gevent
    #   sqlalchemy
gtfparse==1.2.1
    # via -r requirements.in
gunicorn==20.0.4
    # via -r requirements.in
idna==2.9
    # via
    #   -r requirements.in
    #   requests
itsdangerous==1.1.0
    # via
    #   -r requirements.in
    #   flask
jinja2==2.11.1
    # via
    #   -r requirements.in
    #   flask
joblib==0.14.1
    # via
    #   -r requirements.in
    #   mlxtend
    #   scikit-learn
kiwisolver==1.1.0
    # via matplotlib
markupsafe==1.1.1
    # via
    #   -r requirements.in
    #   jinja2
marshmallow==3.5.1
    # via
    #   -r requirements.in
    #   flask-marshmallow
matplotlib==3.2.1
    # via
    #   -r requirements.in
    #   mlxtend
mlxtend==0.17.3
    # via -r requirements.in
networkx==2.5
    # via obonet
numpy==1.22.1
    # via
    #   -r requirements.in
    #   gtfparse
    #   matplotlib
    #   mlxtend
    #   pandas
    #   patsy
    #   pygam
    #   scikit-learn
    #   scipy
    #   statsmodels
obonet==0.2.6
    # via -r requirements.in
packaging==20.3
    # via -r requirements.in
pandas==1.4.0
    # via
    #   -r requirements.in
    #   gtfparse
    #   mlxtend
    #   pyreadr
    #   statsmodels
parso==0.7.1
    # via -r requirements.in
patsy==0.5.2
    # via statsmodels
pillow==7.0.0
    # via -r requirements.in
pip-tools==6.5.1
    # via -r requirements.in
progressbar2==4.0.0
    # via pygam
psutil==5.7.0
    # via -r requirements.in
psycopg2-binary==2.8.4
    # via -r requirements.in
ptyprocess==0.6.0
    # via -r requirements.in
pygam==0.8.0
    # via -r requirements.in
pyparsing==2.4.6
    # via
    #   -r requirements.in
    #   matplotlib
    #   packaging
pyreadr==0.4.4
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   matplotlib
    #   pandas
python-dotenv==0.12.0
    # via -r requirements.in
python-magic==0.4.18
    # via -r requirements.in
python-utils==3.1.0
    # via progressbar2
pytz==2021.3
    # via
    #   -r requirements.in
    #   flask-restful
    #   pandas
pyyaml==5.3.1
    # via -r requirements.in
requests==2.23.0
    # via -r requirements.in
scikit-learn==0.22.2.post1
    # via
    #   -r requirements.in
    #   mlxtend
    #   sklearn
scipy==1.8.0
    # via
    #   -r requirements.in
    #   mlxtend
    #   pygam
    #   scikit-learn
    #   statsmodels
six==1.14.0
    # via
    #   -r requirements.in
    #   cycler
    #   flask-cors
    #   flask-marshmallow
    #   flask-restful
    #   packaging
    #   patsy
    #   pip-tools
    #   python-dateutil
sklearn==0.0
    # via -r requirements.in
sqlalchemy==1.4.32
    # via -r requirements.in
statsmodels==0.13.1
    # via -r requirements.in
urllib3==1.25.8
    # via
    #   -r requirements.in
    #   requests
wcwidth==0.1.9
    # via -r requirements.in
werkzeug==1.0.0
    # via
    #   -r requirements.in
    #   flask
wheel==0.37.1
    # via -r requirements.in
xlrd==1.2.0
    # via -r requirements.in
zope-event==4.4
    # via gevent
zope-interface==5.1.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
